style: Переименованы сервисы обработки сообщений, классы перенесены в соответствующие пакеты + рефакторинг под новые условия
