+++ src/main/java/com/tcs/edu/service/ProcessingMessageService.java
@@ -1,92 +1,39 @@
 package com.tcs.edu.service;
 
 import com.tcs.edu.domain.LogException;
-import com.tcs.edu.enums.Doubling;
-import com.tcs.edu.enums.MessageOrder;
-import com.tcs.edu.interfaces.MessageDecorator;
-import com.tcs.edu.interfaces.MessageService;
 import com.tcs.edu.domain.Message;
-import com.tcs.edu.interfaces.Printer;
+import com.tcs.edu.enums.Severity;
+import com.tcs.edu.interfaces.ProcessingMessageServiceInterface;
+import com.tcs.edu.repository.HashMapMessageRepository;
+import com.tcs.edu.interfaces.MessageRepository;
 
-import static com.tcs.edu.decorator.CutDecorator.*;
-import static java.util.Arrays.stream;
+import java.util.Collection;
+import java.util.UUID;
 
-/**
- * <p>Класс для конкатенации и отправки в принтер отдекорированных сообщений<p/>
- *
- * @author Каримов Ришат
- */
-public class ProcessingMessageService extends ValidatedService implements MessageService {
+public class ProcessingMessageService extends ValidatedService implements ProcessingMessageServiceInterface {
+    private MessageRepository messageRepository = new HashMapMessageRepository();
 
-    private static final MessageOrder DEFAULT_ORDER = MessageOrder.ASC;
-
-    private Printer printer;
-    private MessageDecorator decorator;
-
-    /**
-     *
-     * Конструктор, принимающий параметры принтера и декоратора
-     *
-     */
-    public ProcessingMessageService(Printer printer, MessageDecorator decorator) {
-        this.printer = printer;
-        this.decorator = decorator;
+    public UUID log(Message message) {
+        try {
+            super.isArgsValid(message);
+        } catch (IllegalArgumentException e) {
+            throw new LogException(e);
+        }
+        return messageRepository.create(message);
     }
 
-    /**
-     * <p>Метод конкатенирует и передает в принтер отдекорированные сообщения<p/>
-     *
-     * @param - print - отдекорированное сообщение со строковым типом
-     */
-    public void log(Message... messages) throws LogException {
-        log(DEFAULT_ORDER, messages);
+    @Override
+    public Message findByPrimaryKey(UUID key) {
+        return messageRepository.findByPrimaryKey(key);
     }
 
-    /**
-     * Метод сортирует массив сообщений в обратном порядке
-     *
-     * @param order    - порядковое значение в массиве
-     * @param messages - входные данные
-     */
-    public void log(MessageOrder order, Message... messages) throws LogException {
-        Message[] heap = new Message[messages.length];
-        if (order == MessageOrder.DESC) {
-
-            for (int count = messages.length - 1; count >= 0; count--) {
-                heap[count] = messages[messages.length - 1 - count];
-            }
-        }
-        else {
-            heap = messages;
-        }
-
-        for (Message m : heap) {
-            try {
-                super.isArgsValid(m);
-                printer.print(cutter(decorator.decorate(m)));
-            }
-            catch (IllegalArgumentException e) {
-                throw new LogException("Can`t print message:" + e.getMessage(), e);
-            }
-        }
+    @Override
+    public Collection<Message> findAll() {
+        return messageRepository.findAll();
     }
 
-    /**
-     * Метод осуществляет проверку на дубли, игнорируя их при выводе
-     *
-     * @param order-    порядковое значение в массиве
-     * @param doubling- параметр дублирования
-     * @param messages- входные данные
-     */
-    public void log(MessageOrder order, Doubling doubling, Message... messages) throws LogException {
-        var doublingType = stream(messages);
-
-        if (doubling == Doubling.DISTINCT) {
-            doublingType = doublingType.distinct();
-        }
-        log(order, doublingType.toArray(Message[]::new));
-
+    @Override
+    public Collection<Message> findBySeverity(Severity severity) {
+        return messageRepository.findBySeverity(severity);
     }
-
 }
-
